+++ android/src/main/kotlin/com/tusaamf/flutter_datalogic/FlutterDatalogicPlugin.kt
@@ -2,7 +2,9 @@ package com.tusaamf.flutter_datalogic
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.content.Intent
 import android.content.IntentFilter
+import android.os.Bundle
 import android.util.Log
 import com.tusaamf.flutter_datalogic.const.MyChannels
 
@@ -14,14 +16,21 @@ import io.flutter.plugin.common.MethodChannel.Result
 import java.lang.Exception
 import com.datalogic.decode.BarcodeManager
 import com.datalogic.decode.DecodeException
+import com.datalogic.decode.StartListener
+import com.datalogic.decode.StopListener
 import com.datalogic.decode.configuration.IntentDeliveryMode
 import com.datalogic.decode.configuration.ScannerProperties
 import com.datalogic.device.configuration.ConfigException
+import com.tusaamf.flutter_datalogic.const.MyMethods
+import com.tusaamf.flutter_datalogic.const.ScannerStatus
+import io.flutter.plugin.common.MethodChannel
 
 /** FlutterDatalogicPlugin */
 class FlutterDatalogicPlugin : FlutterPlugin, MethodCallHandler, EventChannel.StreamHandler {
 
     private lateinit var scanEventChannel: EventChannel
+
+    private lateinit var commandMethodChannel: MethodChannel
 
     private var manager: BarcodeManager? = null
 
@@ -37,9 +46,12 @@ class FlutterDatalogicPlugin : FlutterPlugin, MethodCallHandler, EventChannel.St
     private lateinit var intentFilter: IntentFilter
 
     override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+        context = flutterPluginBinding.applicationContext
+
         try {
             // Create a BarcodeManager. It will be used later to change intent delivery modes.
             manager = BarcodeManager()
+
             // get the current settings from the BarcodeManager
             configuration = ScannerProperties.edit(manager)
 
@@ -66,16 +78,20 @@ class FlutterDatalogicPlugin : FlutterPlugin, MethodCallHandler, EventChannel.St
         }
 
         try {
-            context = flutterPluginBinding.applicationContext
-
             // Register dynamically decode wedge intent broadcast receiver.
             intentFilter = IntentFilter()
+            intentFilter.addAction(DLInterface.ACTION_SCANNER_STATUS)
             intentFilter.addAction(DLInterface.ACTION_BROADCAST_RECEIVER)
             intentFilter.addCategory(DLInterface.CATEGORY_BROADCAST_RECEIVER)
 
             scanEventChannel =
                 EventChannel(flutterPluginBinding.binaryMessenger, MyChannels.scanChannel)
             scanEventChannel.setStreamHandler(this)
+
+            commandMethodChannel =
+                MethodChannel(flutterPluginBinding.binaryMessenger, MyChannels.commandChannel)
+            commandMethodChannel.setMethodCallHandler(this)
+
         } catch (e: Exception) {
             Log.e(LOG_TAG, "Error while register intent broadcast receiver", e)
             e.printStackTrace()
@@ -83,6 +99,43 @@ class FlutterDatalogicPlugin : FlutterPlugin, MethodCallHandler, EventChannel.St
     }
 
     override fun onMethodCall(call: MethodCall, result: Result) {
+        when (call.method) {
+            MyMethods.listenScannerStatus -> {
+                listenScannerStatus()
+                result.success(null)  //  Datalogic does not return responses
+            }
+
+            else -> {
+                result.notImplemented()
+            }
+        }
+    }
+
+    private fun listenScannerStatus() {
+        val startListener = StartListener {
+            Intent().also { intent ->
+                val bundle = Bundle().also {
+                  it.putString(DLInterface.EXTRA_KEY_VALUE_SCANNER_STATUS, ScannerStatus.SCANNING.toString())
+                }
+
+                intent.action = DLInterface.ACTION_SCANNER_STATUS
+                intent.putExtra(DLInterface.EXTRA_SCANNER_STATUS, bundle)
+                context.sendBroadcast(intent)
+            }
+        }
+        val stopListener = StopListener {
+            Intent().also { intent ->
+                val bundle = Bundle().also {
+                    it.putString(DLInterface.EXTRA_KEY_VALUE_SCANNER_STATUS, ScannerStatus.IDLE.toString())
+                }
+
+                intent.action = DLInterface.ACTION_SCANNER_STATUS
+                intent.putExtra(DLInterface.EXTRA_SCANNER_STATUS, bundle)
+                context.sendBroadcast(intent)
+            }
+        }
+        manager?.addStartListener(startListener)
+        manager?.addStopListener(stopListener)
     }
 
     override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {