+++ android/src/main/kotlin/com/tusaamf/flutter_datalogic/SinkBroadcastReceiver.kt
@@ -4,6 +4,7 @@ import android.content.BroadcastReceiver
 import android.content.ContentValues.TAG
 import android.content.Context
 import android.content.Intent
+import android.os.Bundle
 import android.util.Log
 import com.tusaamf.flutter_datalogic.const.MyEvents
 import io.flutter.plugin.common.EventChannel.EventSink
@@ -18,8 +19,15 @@ class SinkBroadcastReceiver(private var events: EventSink? = null) : BroadcastRe
                 handleScannedBarcode(intent)
             }
 
+            DLInterface.ACTION_SCANNER_STATUS -> {
+                if (intent.hasExtra(DLInterface.EXTRA_SCANNER_STATUS)) {
+                    intent.getBundleExtra(DLInterface.EXTRA_SCANNER_STATUS)?.let {
+                        handleScannerStatus(it)
+                    }
+                }
+            }
+
             else -> {
-                // Log.d("flutter_datalogic:onReceive:default", intentToString(intent))
                 Log.d(TAG, "default_case")
             }
         }
@@ -31,6 +39,16 @@ class SinkBroadcastReceiver(private var events: EventSink? = null) : BroadcastRe
         val scanResult = JSONObject()
         scanResult.put(MyEvents.EVENT_NAME, MyEvents.SCAN_RESULT)
         scanResult.put("scanData", scanData)
+
+        events!!.success(scanResult.toString())
+    }
+
+    private fun handleScannerStatus(b: Bundle) {
+        val status =
+            b.getString(DLInterface.EXTRA_KEY_VALUE_SCANNER_STATUS)
+        val scanResult = JSONObject()
+        scanResult.put(MyEvents.EVENT_NAME, MyEvents.SCANNER_STATUS)
+        scanResult.put("status", status)
 
         events!!.success(scanResult.toString())
     }