+++ android/src/main/kotlin/com/tusaamf/flutter_datalogic/const/MyChannels.kt
@@ -2,6 +2,7 @@ package com.tusaamf.flutter_datalogic.const
 
 class MyChannels {
     companion object {
+        const val commandChannel = "channels/datalogic_command"
         const val scanChannel = "channels/datalogic_scan"
     }
 }