+++ android/src/main/kotlin/com/tusaamf/flutter_datalogic/const/MyEvents.kt
@@ -3,6 +3,7 @@ package com.tusaamf.flutter_datalogic.const
 class MyEvents {
     companion object {
         const val EVENT_NAME = "EVENT_NAME"
+        const val SCANNER_STATUS = "SCANNER_STATUS"
         const val SCAN_RESULT = "SCAN_RESULT"
     }
 }