+++ android/src/main/kotlin/com/tusaamf/flutter_datalogic/const/MyMethods.kt
@@ -1,0 +1,7 @@
+package com.tusaamf.flutter_datalogic.const
+
+class MyMethods {
+    companion object {
+        const val listenScannerStatus = "listenScannerStatus"
+    }
+}