+++ android/src/main/kotlin/com/tusaamf/flutter_datalogic/const/ScannerStatus.kt
@@ -1,0 +1,6 @@
+package com.tusaamf.flutter_datalogic.const
+
+enum class ScannerStatus {
+    SCANNING,
+    IDLE
+}