+++ lib/flutter_datalogic.dart
@@ -1,1 +1,4 @@
+export 'src/models/scanner_status.dart';
+export 'src/models/scan_result.dart';
+
 export 'src/flutter_datalogic.dart';