+++ lib/src/consts/datalogic_event_type.dart
@@ -1,4 +1,5 @@
 enum DatalogicEventType {
+  scannerStatus('SCANNER_STATUS'),
   scanResult('SCAN_RESULT');
 
   final String value;