+++ lib/src/consts/method_channel_methods.dart
@@ -1,0 +1,7 @@
+enum MethodChannelMethods {
+  listenScannerStatus("listenScannerStatus");
+
+  final String value;
+
+  const MethodChannelMethods(this.value);
+}