+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Changelog
 
-## [1.0.1] - 27th October 2023
+## [1.0.3] - 23th November 2023
+
+* Add listen Scanner status
+
+## [1.0.2] - 22th November 2023
+
+* Update build settings
+
+## [1.0.1] - 22th November 2023
 
 * Update sample on README.md 
 
-## [1.0.0] - 27th October 2023
+## [1.0.0] - 22th November 2023
 
 * Initial release